requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
openai>=1.10.0
tiktoken>=0.5.2

# Vector databases and search
chromadb>=0.4.22
faiss-cpu>=1.7.4
pinecone-client>=3.0.0

# Graph processing (for Graph RAG)
networkx>=3.2.1
neo4j>=5.15.0
pyvis>=0.3.2

# Data processing
pandas>=2.1.4
numpy>=1.24.3
scikit-learn>=1.3.2

# Text processing and embeddings
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.2

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.1
loguru>=0.7.2

# API and web
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.29.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1